+++ eventmesh-connector-api/src/main/java/com/webank/eventmesh/api/AbstractContext.java
@@ -1,0 +1,4 @@
+package com.webank.eventmesh.api;
+
+public interface AbstractContext {
+}
